re-added missed changes from mesh serialization feature branch; fixed ios build

+++ src/moai-util/MOAIStream.h
@@ -142,7 +142,7 @@ private:
 				}
 			}
 		}
-		state.Push ( bytes );
+		state.Push (( u32 )bytes ); // TODO: overflow?
 		return 1;
 	}
